ref: moved delegate into own file
