+++ src/flow_queue/mod.rs
@@ -1,3 +1,4 @@
 pub mod connection;
 pub mod name;
 pub mod handler;
+pub mod delegate;
